requirements.txt
pandas>=2.0
openpyxl>=3.1
matplotlib>=3.8
seaborn>=0.13
numpy>=1.26
python-pptx>=0.6.23
python-docx>=1.1.0
google-genai>=1.55.0
latex2mathml>=3.77
Pillow>=10.0
lxml>=5.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
jinja2>=3.1.0
markdown>=3.6
